Updating to new app samples now only one for VS and one for Scalar - VS code.
