--- aspnetcore/tutorials/min-web-api/samples/10.x/todo/Message.cs
@@ -1,8 +1,0 @@
-public class Message
-{
-    public Message()
-    {
-    }
-
-    public string Text { get; set; }
-}
