+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo/TodoApi/Todo.cs
@@ -3,4 +3,5 @@
     public int Id { get; set; }
     public string? Name { get; set; }
     public bool IsComplete { get; set; }
+    public string? Secret { get; set; }
 }
