+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo/TodoApi/TodoApi.csproj
