+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo/TodoApi/TodoItemDTO.cs
@@ -1,4 +1,4 @@
-public class TodoItemDTO
+﻿public class TodoItemDTO
 {
     public int Id { get; set; }
     public string? Name { get; set; }
