+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo/TodoApi/TodoPatchDto.cs
