--- aspnetcore/tutorials/min-web-api/samples/10.x/todoDTO/Todo.cs
@@ -1,7 +1,0 @@
-public class Todo
-{
-    public int Id { get; set; }
-    public string? Name { get; set; }
-    public bool IsComplete { get; set; }
-    public string? Secret { get; set; }
-}
