--- aspnetcore/tutorials/min-web-api/samples/10.x/todoGroup/Todo.cs
@@ -1,6 +1,0 @@
-public class Todo
-{
-    public int Id { get; set; }
-    public string? Name { get; set; }
-    public bool IsComplete { get; set; }
-}
