+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo_ScalarVersion/TodoApi/Todo.cs
@@ -1,6 +1,7 @@
-class Todo
+﻿public class Todo
 {
     public int Id { get; set; }
     public string? Name { get; set; }
     public bool IsComplete { get; set; }
+    public string? Secret { get; set; }
 }
