+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo_ScalarVersion/TodoApi/TodoDb.cs
@@ -1,4 +1,4 @@
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 
 class TodoDb : DbContext
 {
