+++ aspnetcore/tutorials/min-web-api/samples/10.x/todo_ScalarVersion/TodoApi/TodoPatchDto.cs
@@ -1,0 +1,5 @@
+﻿public class TodoPatchDto
+{
+    public string? Name { get; set; }
+    public bool? IsComplete { get; set; }
+}
