--- aspnetcore/tutorials/min-web-api/samples/10.x/todo_SwaggerVersion/TodoDb.cs
@@ -1,9 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-
-class TodoDb : DbContext
-{
-    public TodoDb(DbContextOptions<TodoDb> options)
-        : base(options) { }
-
-    public DbSet<Todo> Todos => Set<Todo>();
-}
